3799: Extract CsgTest
