+++ common/test/CMakeLists.txt
@@ -94,6 +94,7 @@ set(COMMON_TEST_SOURCE
         "${COMMON_TEST_SOURCE_DIR}/View/CommandProcessorTest.cpp"
         "${COMMON_TEST_SOURCE_DIR}/View/CompilationRunToolTaskRunnerTest.cpp"
         "${COMMON_TEST_SOURCE_DIR}/View/CopyPasteTest.cpp"
+        "${COMMON_TEST_SOURCE_DIR}/View/CsgTest.cpp"
         "${COMMON_TEST_SOURCE_DIR}/View/GridTest.cpp"
         "${COMMON_TEST_SOURCE_DIR}/View/GroupNodesTest.cpp"
         "${COMMON_TEST_SOURCE_DIR}/View/InputEventTest.cpp"
